refactor: :truck: firebaseディレクトリをlibsディレクトリ以下に移動
