+++ packages/app/src/shared/libs/firebase/config.ts
